pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zoom-midi-host"
version = "0.1.0"
description = "Raspberry Pi host application for controlling Zoom MS-60B+ via MIDI"
authors = [
    {name = "Vibe Coding", email = ""}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux"
]
dependencies = [
    "mido>=1.3.0",
    "python-rtmidi>=1.5.8",
    "pyusb>=1.2.1",
    "Pillow>=9.5.0",
    "luma.lcd>=2.6.0",
    "dataclasses-json>=0.6.3"
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "ruff>=0.1.7",
    "mypy>=1.6"
]

[project.scripts]
zoom-midi-host = "zoom_midi_host.__main__:main"
